# src/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Project 1 API",
    "description": "Send forward your gaming community using Project 1 API",
    "version": "1.0.0"
  },
  "produces": [
    "application/json"
  ],
...
  "basePath": "/api",
  "paths": {
...
    "/playerAccount": {
      "get": {
        "x-swagger-router-controller": "PlayerAccountController",
        "operationId": "getPlayerAccountList",
        "tags": [
          "PlayerAccount"
        ],
        "summary": "Get all playerAccount",
        "description": "Gets `playerAccount` objects.\nOptional query param of **size** determines\nsize of returned array\n",
        "parameters": [
          {
            "name": "size",
            "in": "query",
            "description": "Size of array",
            "required": false,
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "An array of PlayerAccount",
            "schema": {
              "title": "ArrayOfPlayerAccount",
              "type": "array",
              "items": {
                "$ref": "#/definitions/PlayerAccount"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/playerAccount/{userId}/addPlayerAccount/{gameId}": {
      "post": {
        "x-swagger-router-controller": "PlayerAccountController",
        "operationId": "addPlayerAccount",
        "tags": [
          "PlayerAccount"
        ],
        "summary": "Adds a new PlayerAccount",
        "description": "Adds new `PlayerAccount` object to the DB.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId of player object to add.",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameId",
            "in": "path",
            "description": "gameId of player object to add.",
            "required": true,
            "type": "string"
          },
          {
            "name": "PlayerAccount",
            "in": "body",
            "description": "PlayerAccount object to insert",
            "required": false,
            "schema": {
              "$ref": "#/definitions/CreatePlayedAccountObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PlayerAccount has been successfully added to the DB"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/playerAccount/{playerAccountId}/getPlayerAccountById": {
      "get": {
        "x-swagger-router-controller": "PlayerAccountController",
        "operationId": "getPlayerAccountById",
        "tags": [
          "PlayerAccount"
        ],
        "summary": "Get the specified playerAccount",
        "description": "Gets `playerAcount` object specified by its ID.",
        "parameters": [
          {
            "name": "playerAccountId",
            "in": "path",
            "description": "playerAcountId of player object to search for.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved playerAcount object",
            "schema": {
              "$ref": "#/definitions/PlayerAccount"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/playerAccount/{userId}/getPlayerAccountByUserId": {
      "get": {
        "x-swagger-router-controller": "PlayerAccountController",
        "operationId": "getPlayerAccountByUserId",
        "tags": [
          "PlayerAccount"
        ],
        "summary": "Get the corresponding playerAccount based on the PlayerAccount UserId",
        "description": "Gets `player` object specified by its UserId.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId of Player object to find.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "An array of Players",
            "schema": {
              "title": "ArrayOfPlayers",
              "type": "array",
              "items": {
                "$ref": "#/definitions/PlayerAccount"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/playerAccount/{login}/getPlayerAccountByLogin": {
      "get": {
        "x-swagger-router-controller": "PlayerAccountController",
        "operationId": "getPlayerAccountByLogin",
        "tags": [
          "PlayerAccount"
        ],
        "description": "Gets `player` object specified by its login.",
        "summary": "Get the corresponding playerAccount based on the PlayerAccount login",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "description": "login of Player object to find.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved PlayerAccount object",
            "schema": {
              "$ref": "#/definitions/PlayerAccount"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/playerAccount/{playerAccountId}/deletePlayerAccount": {
      "put": {
        "x-swagger-router-controller": "PlayerAccountController",
        "operationId": "deletePlayerAccount",
        "tags": [
          "PlayerAccount"
        ],
        "description": "Deactivates the 'playerAccount' object specified by its ID.",
        "summary": "Deactivate the corresponding playerAccount",
        "parameters": [
          {
            "name": "playerAccountId",
            "in": "path",
            "description": "playerAccountId of Game object to deactivate.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated playerAccount object",
            "schema": {
              "$ref": "#/definitions/PlayerAccount"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "birthDate": {
          "type": "string",
          "format": "date"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "format": "password"
        },
        "avatar": {
          "type": "string"
        },
        "address": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "phoneNumber": {
          "type": "string"
        },
        "admin": {
          "type": "boolean"
        },
        "friends": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        },
        "active": {
          "type": "boolean"
        },
        "created_at": {
          "type": "string",
          "format": "date"
        },
        "updated_at": {
          "type": "string",
          "format": "date"
        }
      }
    },
...
    "PlayerAccount": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        },
        "login": {
          "type": "string"
        },
        "game": {
          "$ref": "#/definitions/Game"
        },
        "active": {
          "type": "boolean"
        },
        "created_at": {
          "type": "string",
          "format": "date"
        },
        "updated_at": {
          "type": "string",
          "format": "date"
        }
      }
    },
    "Address": {
      "type": "object",
      "properties": {
        "postCode": {
          "type": "integer",
          "format": "int32"
        },
        "city": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "line": {
          "type": "string"
        }
      }
    },
    "Game": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "releaseDate": {
          "type": "string",
          "format": "date"
        },
        "multiPlayer": {
          "type": "boolean"
        },
        "editor": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "date"
        },
        "updated_at": {
          "type": "string",
          "format": "date"
        }
      }
    },
...
    "CreatePlayedAccountObject": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string"
        }
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "fields": {
          "type": "string"
        }
      }
    }
  }
}